requirements.txt
coreapi==2.3.3
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1
django-cors-headers==3.0.1

